pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code-orbit"
version = "0.1.0"
description = "AI-powered codebase editor using local llama.cpp models"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
llm-agent = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]
